pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.3",
    "httpx>=0.28.1",
    "plaid-python>=38.1.0",
    "sse-starlette>=3.2.0",
    "uvicorn>=0.40.0",
    "google-genai>=0.8.0",
    "google-api-python-client>=2.138.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
]
